+++ packages/upload/index.ts
@@ -1,0 +1,6 @@
+/* eslint-disable */
+export * from './upload.inject';
+export * from './upload.module';
+export * from './upload.service';
+export * from './upload.types';
+export * from './upload.config';