+++ packages/upload/upload.inject.ts
@@ -1,0 +1,5 @@
+import { Inject } from '@nestjs/common';
+
+export const UPLOAD_CONFIG_KEY = 'UPLOAD_CONFIG_KEY';
+
+export const InjectUploadConfig = Inject(UPLOAD_CONFIG_KEY);